Set up proper targets for copied files in Makefile
When building the extension, we need to copy a bunch of different
source files into the build directory. Setting up targets for each
copied file manually would be difficult and error-prone, so until now
we used an empty target combined with rsync as a workaround.

Instead of that workaround, let's define helper functions that can
be used to set up proper targets (using static pattern rules[1]) for
the copied files. That way, each file can be copied individually
should it be modified.

This also means that copied file targets can be responsible for
creating their own directory structure, so we can drop the
MKDIR_TARGETS step.

1 - https://www.gnu.org/software/make/manual/html_node/Static-Usage.html